% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonview-package.r
\docType{package}
\name{jsonview}
\alias{jsonview}
\alias{jsonview-package}
\title{Format and Pretty Print JSON}
\description{
This uses \code{vkbeautify} and \code{highlight.js} to format and
"pretty print" JSON content & lists in an \code{htmlwidget} pane.
}
\details{
Also uses \code{jquery.json-view} for the interactive tree view
}
\author{
Bob Rudis (@hrbrmstr)
}

