% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jview.R
\name{jview}
\alias{jview}
\title{jview JSON viewer}
\usage{
jview(data, width = NULL, height = NULL, elementId = NULL, expanded = FALSE)
}
\arguments{
\item{data}{A json string}

\item{width}{widget width}

\item{height}{widget height}

\item{elementId}{shiny ID}

\item{expanded}{Boolean.  If true then start in expanded form}
}
\description{
Display a JSON string as a collapsible tree
}
\examples{
if (interactive()) {
  library(shiny)
  library(jsonview)

  json_test <- '{
    "givenName": "Vas",
    "surName": "Sudanagunta",
    "children": [
      {
        "givenName": "Natalia",
        "age": 5
      },
      {
        "givenName": "Aida",
        "age": 17
      }
    ],
    "address": {
      "state": "NY",
      "city": "Brooklyn",
      "street": "718 Marcus Garvey Ave"
    }
  }'

  shiny::shinyApp(
    ui = shiny::fluidPage(
      jsonview::jviewOutput("json")
    ),
    server = function(input, output) {
      output$json <- jsonview::renderJview({
        jsonview::jview(data = json_test, expanded = TRUE)
      })
    }
  )
}

}
