% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonview.r
\name{json_view}
\alias{json_view}
\title{JSON viewer}
\usage{
json_view(x, auto_unbox = TRUE, style = "default", scroll = FALSE,
  elementId = NULL, width = "100\%", height = NULL)
}
\description{
JSON viewer
}

