% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jview.R
\name{jview}
\alias{jview}
\title{jview JSON viewer}
\usage{
jview(data, width = NULL, height = NULL, elementId = NULL, expanded = FALSE)
}
\arguments{
\item{data}{A json string}

\item{width}{widget width}

\item{height}{widget height}

\item{elementId}{shiny ID}

\item{expanded}{Boolean.  If true then start in expanded form}
}
\description{
Display a JSON string as a collapsible tree
}
